Add files via upload
Hello, thanks for adding this printer to example configs a while back. I gave you an old version and you made it up to date and i tested it with bugfix version.
I configurated this printer and added useful functions that are 100% compatible assuming a stock printer. I made it the best of my abilities and i think thats the best marlin configuration for that printer. I put in much effort :)
However, i am not a programmer, just doint that as a hobby. So i am afraid that i am not able or lets say i will mess it up if i try to get it up to date.

so i hope you can please do the magic and use my configs somehow so everyone with stock Fabrikator will get the most out of it :) but everything else is like i said too much for me unfortunatly and i really dont want to mess up anything. 

Thank you again for adding this printer in the first place <3

i hope i did that correctly with submiting its the first time for me doing pull request.
Like i said this is from the bugfix version that i tested and it need to be updated. this config is already some weeks/month old and i saw that in the meantime something was changed with tmc stepper i dont understand. so please take it in consideration that my configs are uploaded only to provide best functionality for printer :)